+++ CHANGELOG.md
@@ -10,6 +10,8 @@ For a roadmap including expected timeline, please refer to [ROADMAP.md](./ROADMA
 
 ## [unreleased]
 
+- BREAKING: spec-toolkit no longer part of the csn-interop-specification package exports
+
 ## [0.4.0]
 
 ### Added